requirements.txt
mysql-connector-python-rf
mysql-connector
pyglet
black
flask
